avancement des classes manager et UserManager.cs

+++ SearchBars/MetierSearchBars/Bar.cs
@@ -11,11 +11,12 @@ namespace MetierSearchBars
         private Dictionary<User, Avis> commentaires;
         private List<Boisson> listBoissons;
         public string Ville { get; set; }
-        public int NoteMoyenne { get; set; }
+        public float NoteMoyenne { get; set; }
         public CoordonneesGPS GPS{ get; set; }
         public bool Restauration { get; set; }
 
-        public Bar(CoordonneesGPS gps, bool restauration = false, int note = -1, string ville="") {
+        public Bar(CoordonneesGPS gps, bool restauration = false, float note = -1, string ville="") 
+        {
             GPS = gps;
             Restauration = restauration;
             NoteMoyenne = note;
